backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database & ORM
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.0

# LangGraph & LLM
langgraph==0.2.0
langchain==0.3.0
langchain-google-genai==2.0.0
google-generativeai==0.8.0

# Search
tavily-python==0.5.0

# Scraping
playwright==1.48.0
feedparser==6.0.11
httpx==0.27.0

# Scheduling
apscheduler==3.10.4

# Utils
python-dotenv==1.0.1
